feat: add functionality for search dropdown component
- This commit will add the functionality for the search dropdown based on click event or blur event
